Move isWhitelisted function to new file
