+++ server/services/email-campaigns/isWhitelisted.js
@@ -1,0 +1,13 @@
+const config = require('../../config/config');
+
+const isWhitelisted = (email) => {
+  if (!config.get('sendInBlue.whitelist')) {
+    return true;
+  }
+
+  return config.get('sendInBlue.whitelist').split(',').includes(email);
+};
+
+module.exports = {
+  isWhitelisted,
+};
